refactor im add * extract AvailableInputMethodView out * add padding to only show current language * move buttons to the right * show import table button only when chinese-addons is installed * auto add imported table to current group
